[FE] refactor: переименовал файлы согласно критерию Б36;
