+++ src/components/offer-card/offer-card.e2e.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {mount} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import OfferCard from "./OfferCard";
+import OfferCard from "./offer-card";
 import {getTestCard} from "../../mocks/get-test-card";
 
 Enzyme.configure({adapter: new Adapter()});
