+++ src/components/offer-details/offer-details.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import {getTestCard} from "../../mocks/get-test-card";
-import OfferDetails from "./OfferDetails";
+import OfferDetails from "./offer-details";
 import {leaflet} from "../../mocks/leaflet";
-import Map from "../map/Map";
+import Map from "../map/map";
 
 it(`OfferDetails correctly renders after relaunch`, () => {
   const card = getTestCard(`OfferDetails snap 2`, 999, false);
