+++ src/components/offer-list/offer-list.jsx
@@ -1,6 +1,6 @@
 import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
-import OfferCard from "../offer-card/OfferCard";
+import OfferCard from "../offer-card/offer-card";
 
 class OfferList extends PureComponent {
   constructor(props) {
