+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import leaflet from 'leaflet';
 import {offers} from "./mocks/offers";
 import {nearbyOffers} from "./mocks/nearby-offers";
-import App from "./components/app/App";
+import App from "./components/app/app";
 
 const init = () => {
